+++ App.vue
@@ -25,6 +25,6 @@
 
 <style lang="scss">
 	/*每个页面公共css */
-	@import "uview-ui/index.scss";
+	// @import "./uni_modules/vk-uview-ui/index.scss";
   @import "./common/css/app.scss";
 </style>