+++ main.js
@@ -3,7 +3,7 @@ import store from './store'
 import config from '@/app.config.js'
 
 // 引入 uView UI
-import uView from 'uview-ui';
+// import uView from './uni_modules/vk-uview-ui';
 // 引入 vk框架前端
 import vk from './uni_modules/vk-unicloud';
 
@@ -11,7 +11,7 @@ import vk from './uni_modules/vk-unicloud';
 import Vue from 'vue'
 
 // 引入 uView UI
-Vue.use(uView);
+// Vue.use(uView);
 
 // 引入 vk框架前端
 Vue.use(vk);
@@ -44,7 +44,7 @@ export function createApp() {
   app.use(store)
   
   // 引入 uView UI
-  app.use(uView)
+  // app.use(uView)
   
   // 引入 vk框架前端
   app.use(vk);